Experiment with new mixin class
